pyproject.toml
[tool.poetry]
name = "sadko-dostavka"
version = "0.1.0"
description = ""
authors = ["Rasul Cheerchiev <rasuliomusic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.3"
Django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg = "^3.1.9"
psycopg-binary = "^3.1.9"
django-admin-sortable2 = "^2.1.8"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pycodestyle = "^2.10.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
